add lazy loading to home page
